pyproject.toml
# =============================================================================
# LocalSingleCell Project Configuration
# 本地化单细胞&空间转录组分析工具
# =============================================================================

[project]
name = "localsinglecell"
version = "3.1.0"
description = "本地化单细胞&空间转录组分析工具 - 完全本地化、无云端依赖"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "LocalSingleCell Team"}
]
keywords = [
    "single-cell",
    "rna-seq",
    "spatial-transcriptomics",
    "bioinformatics",
    "scanpy",
    "streamlit"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Streamlit",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

dependencies = [
    "streamlit>=1.32.2",
    "scanpy>=1.10.1",
    "squidpy>=1.4.1",
    "anndata>=0.10.7",
    "leidenalg>=0.10.2",
    "umap-learn>=0.5.6",
    "numpy>=1.26.4",
    "pandas>=2.2.1",
    "scipy>=1.12.0",
    "scikit-learn>=1.4.1",
    "matplotlib>=3.8.3",
    "seaborn>=0.13.2",
    "plotly>=5.20.0",
    "gseapy>=1.1.2",
    "h5py>=3.10.0",
    "openpyxl>=3.1.2",
    "pyyaml>=6.0.1",
    "psutil>=5.9.8",
    "tqdm>=4.66.2",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "black>=24.1.0",
    "isort>=5.13.0",
]
docs = [
    "sphinx>=7.2.0",
    "sphinx-rtd-theme>=2.0.0",
    "myst-parser>=2.0.0",
]
sra = [
    # SRA Toolkit 需要单独安装
]

[project.urls]
Homepage = "https://github.com/your-repo/LocalSingleCell"
Documentation = "https://github.com/your-repo/LocalSingleCell#readme"
Repository = "https://github.com/your-repo/LocalSingleCell"
Issues = "https://github.com/your-repo/LocalSingleCell/issues"

# =============================================================================
# Ruff 配置 (代码检查和格式化)
# =============================================================================
[tool.ruff]
line-length = 120
target-version = "py310"
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "*.egg-info",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (交给格式化处理)
    "B008",   # function call in default argument
    "B905",   # zip without strict
    "ARG001", # unused function argument
    "SIM108", # ternary operator
]

# 每个文件的忽略规则
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # 允许未使用的导入
"tests/*" = ["ARG"]       # 测试文件允许未使用参数

# =============================================================================
# isort 配置 (导入排序)
# =============================================================================
[tool.isort]
profile = "black"
line_length = 120
known_first_party = ["core", "utils", "ui"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
skip = [".git", ".venv", "venv", "__pycache__"]

# =============================================================================
# pytest 配置
# =============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::FutureWarning",
]

# =============================================================================
# coverage 配置
# =============================================================================
[tool.coverage.run]
source = ["core", "utils"]
branch = true
omit = [
    "*/__init__.py",
    "*/tests/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]

# =============================================================================
# mypy 配置 (类型检查)
# =============================================================================
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = [
    "tests/",
    "venv/",
    ".venv/",
]

# =============================================================================
# black 配置 (代码格式化)
# =============================================================================
[tool.black]
line-length = 120
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | venv
    | __pycache__
    | build
    | dist
)/
'''

# =============================================================================
# 构建系统配置
# =============================================================================
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "utils*", "ui*"]
